Develop

+++ src/main/java/com/cs/jeyz9/condoswiftapi/dto/AnnounceDTO.java
@@ -17,7 +17,7 @@ public class AnnounceDTO {
     private Double price;
     private Integer bathroomCount;
     private Integer bedroomCount;
-    private Integer areaSize;
+    private Double areaSize;
     private Boolean hasPool;
     private Boolean hasConvenienceStore;
     private Boolean hasFitness;
